requirements.txt
# Core dependencies
langgraph>=0.0.1
langchain>=0.1.0
openai>=1.0.0
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0
python-dotenv>=1.0.0

# Phase 3: Delivery & Auditability
pygithub>=2.1.0
requests>=2.31.0

# Phase 4: Organizational Memory & Learning
# No additional dependencies (uses built-in sqlite3)

# Phase 5: Safety, Policy & Governance
bandit>=1.7.5

# Phase 6: Human Interface & Operations
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pyjwt>=2.8.0

# Development
pytest>=7.0.0
